requirements.txt
tweepy==3.6.0
requests==2.19.1
python-decouple==3.1